tar: Add support for creating tgz archives
